refactor: modularized unit and integ testing
This commit introduced modular unit and integ testing via a file
refactor and splitting up 'urban' class into seperate classes into
seperate files.

Authored-by: Joshua Rose <joshuarose099@gmail.com>
